+++ packages/twenty-server/src/modules/calendar/services/calendar-event-participant-manager/calendar-event-participant.service.ts
