+++ packages/twenty-server/src/modules/calendar/services/calendar-sync-manager/google-calendar-sync.service.ts
@@ -22,7 +22,7 @@ import { CalendarChannelEventAssociationWorkspaceEntity } from 'src/modules/cale
 import { CalendarEventParticipantWorkspaceEntity } from 'src/modules/calendar/standard-objects/calendar-event-participant.workspace-entity';
 import { BlocklistWorkspaceEntity } from 'src/modules/connected-account/standard-objects/blocklist.workspace-entity';
 import { CalendarEventCleanerService } from 'src/modules/calendar/services/calendar-event-cleaner/calendar-event-cleaner.service';
-import { CalendarEventParticipantService } from 'src/modules/calendar/services/calendar-event-participant/calendar-event-participant.service';
+import { CalendarEventParticipantService } from 'src/modules/calendar/services/calendar-event-participant-manager/calendar-event-participant.service';
 import {
   CalendarEventParticipant,
   CalendarEventWithParticipants,
